Dissect trial.py such that it doesn't break existing tests
