+++ optuna/trials/trial.py
@@ -1,6 +1,11 @@
+import datetime
+import warnings
+
 from optuna import distributions
 from optuna import logging
 from optuna import type_checking
+from optuna.trials.base import BaseTrial
+from optuna.trials.util import _adjust_discrete_uniform_high
 
 if type_checking.TYPE_CHECKING:
     from typing import Any  # NOQA
@@ -16,359 +21,6 @@ if type_checking.TYPE_CHECKING:
     FloatingPointDistributionType = Union[
         distributions.UniformDistribution, distributions.LogUniformDistribution
     ]
-
-_logger = logging.get_logger(__name__)
-
-
-class TrialState(enum.Enum):
-    """State of a :class:`~optuna.trial.Trial`.
-
-    Attributes:
-        RUNNING:
-            The :class:`~optuna.trial.Trial` is running.
-        COMPLETE:
-            The :class:`~optuna.trial.Trial` has been finished without any error.
-        PRUNED:
-            The :class:`~optuna.trial.Trial` has been pruned with
-            :class:`~optuna.exceptions.TrialPruned`.
-        FAIL:
-            The :class:`~optuna.trial.Trial` has failed due to an uncaught error.
-
-            .. note::
-                The failed trials are ignored by any build-in samplers when they sample new
-                parameters. Thus, failed trials are regarded as deleted in the samplers'
-                perspective.
-    """
-
-    RUNNING = 0
-    COMPLETE = 1
-    PRUNED = 2
-    FAIL = 3
-    WAITING = 4
-
-    def __repr__(self):
-        # type: () -> str
-
-        return str(self)
-
-    def is_finished(self):
-        # type: () -> bool
-
-        return self != TrialState.RUNNING and self != TrialState.WAITING
-
-
-class FrozenTrial(object):
-    """Status and results of a :class:`~optuna.trial.Trial`.
-
-    Attributes:
-        number:
-            Unique and consecutive number of :class:`~optuna.trial.Trial` for each
-            :class:`~optuna.study.Study`. Note that this field uses zero-based numbering.
-        state:
-            :class:`TrialState` of the :class:`~optuna.trial.Trial`.
-        value:
-            Objective value of the :class:`~optuna.trial.Trial`.
-        datetime_start:
-            Datetime where the :class:`~optuna.trial.Trial` started.
-        datetime_complete:
-            Datetime where the :class:`~optuna.trial.Trial` finished.
-        params:
-            Dictionary that contains suggested parameters.
-        user_attrs:
-            Dictionary that contains the attributes of the :class:`~optuna.trial.Trial` set with
-            :func:`optuna.trial.Trial.set_user_attr`.
-        intermediate_values:
-            Intermediate objective values set with :func:`optuna.trial.Trial.report`.
-    """
-
-    def __init__(
-        self,
-        number,  # type: int
-        state,  # type: TrialState
-        value,  # type: Optional[float]
-        datetime_start,  # type: Optional[datetime.datetime]
-        datetime_complete,  # type: Optional[datetime.datetime]
-        params,  # type: Dict[str, Any]
-        distributions,  # type: Dict[str, BaseDistribution]
-        user_attrs,  # type: Dict[str, Any]
-        system_attrs,  # type: Dict[str, Any]
-        intermediate_values,  # type: Dict[int, float]
-        trial_id,  # type: int
-    ):
-        # type: (...) -> None
-
-        self.number = number
-        self.state = state
-        self.value = value
-        self.datetime_start = datetime_start
-        self.datetime_complete = datetime_complete
-        self.params = params
-        self.user_attrs = user_attrs
-        self.system_attrs = system_attrs
-        self.intermediate_values = intermediate_values
-        self._distributions = distributions
-        self._trial_id = trial_id
-
-    # Ordered list of fields required for `__repr__`, `__hash__` and dataframe creation.
-    # TODO(hvy): Remove this list in Python 3.6 as the order of `self.__dict__` is preserved.
-    _ordered_fields = [
-        "number",
-        "value",
-        "datetime_start",
-        "datetime_complete",
-        "params",
-        "_distributions",
-        "user_attrs",
-        "system_attrs",
-        "intermediate_values",
-        "_trial_id",
-        "state",
-    ]
-
-    def __eq__(self, other):
-        # type: (Any) -> bool
-
-        if not isinstance(other, FrozenTrial):
-            return NotImplemented
-        return other.__dict__ == self.__dict__
-
-    def __lt__(self, other):
-        # type: (Any) -> bool
-
-        if not isinstance(other, FrozenTrial):
-            return NotImplemented
-
-        return self.number < other.number
-
-    def __le__(self, other):
-        # type: (Any) -> bool
-
-        if not isinstance(other, FrozenTrial):
-            return NotImplemented
-
-        return self.number <= other.number
-
-    def __hash__(self):
-        # type: () -> int
-
-        return hash(tuple(getattr(self, field) for field in self._ordered_fields))
-
-    def __repr__(self):
-        # type: () -> str
-
-        return "{cls}({kwargs})".format(
-            cls=self.__class__.__name__,
-            kwargs=", ".join(
-                "{field}={value}".format(
-                    field=field if not field.startswith("_") else field[1:],
-                    value=repr(getattr(self, field)),
-                )
-                for field in self._ordered_fields
-            ),
-        )
-
-    def _validate(self):
-        # type: () -> None
-
-        if self.datetime_start is None:
-            raise ValueError("`datetime_start` is supposed to be set.")
-
-        if self.state.is_finished():
-            if self.datetime_complete is None:
-                raise ValueError("`datetime_complete` is supposed to be set for a finished trial.")
-        else:
-            if self.datetime_complete is not None:
-                raise ValueError(
-                    "`datetime_complete` is supposed to be None for an unfinished trial."
-                )
-
-        if self.state == TrialState.COMPLETE and self.value is None:
-            raise ValueError("`value` is supposed to be set for a complete trial.")
-
-        if set(self.params.keys()) != set(self.distributions.keys()):
-            raise ValueError(
-                "Inconsistent parameters {} and distributions {}.".format(
-                    set(self.params.keys()), set(self.distributions.keys())
-                )
-            )
-
-        for param_name, param_value in self.params.items():
-            distribution = self.distributions[param_name]
-
-            param_value_in_internal_repr = distribution.to_internal_repr(param_value)
-            if not distribution._contains(param_value_in_internal_repr):
-                raise ValueError(
-                    "The value {} of parameter '{}' isn't contained in the distribution "
-                    "{}.".format(param_value, param_name, distribution)
-                )
-
-    @property
-    def distributions(self):
-        # type: () -> Dict[str, BaseDistribution]
-        """Dictionary that contains the distributions of :attr:`params`."""
-
-        return self._distributions
-
-    @distributions.setter
-    def distributions(self, value):
-        # type: (Dict[str, BaseDistribution]) -> None
-        self._distributions = value
-
-    @property
-    def trial_id(self):
-        # type: () -> int
-        """Return the trial ID.
-
-        .. deprecated:: 0.19.0
-            The direct use of this attribute is deprecated and it is recommended that you use
-            :attr:`~optuna.trial.FrozenTrial.number` instead.
-
-        Returns:
-            The trial ID.
-        """
-
-        warnings.warn(
-            "The use of `FrozenTrial.trial_id` is deprecated. "
-            "Please use `FrozenTrial.number` instead.",
-            DeprecationWarning,
-        )
-
-        _logger.warning(
-            "The use of `FrozenTrial.trial_id` is deprecated. "
-            "Please use `FrozenTrial.number` instead."
-        )
-
-        return self._trial_id
-
-    @property
-    def last_step(self):
-        # type: () -> Optional[int]
-
-        if len(self.intermediate_values) == 0:
-            return None
-        else:
-            return max(self.intermediate_values.keys())
-
-    @property
-    def duration(self):
-        # type: () -> Optional[datetime.timedelta]
-        """Return the elapsed time taken to complete the trial.
-
-        Returns:
-            The duration.
-        """
-
-        if self.datetime_start and self.datetime_complete:
-            return self.datetime_complete - self.datetime_start
-        else:
-            return None
-
-
-class BaseTrial(object, metaclass=abc.ABCMeta):
-    """Base class for trials.
-
-    Note that this class is not supposed to be directly accessed by library users.
-    """
-
-    @abc.abstractmethod
-    def suggest_float(self, name, low, high, *, log=False, step=None):
-        # type: (str, float, float, bool, Optional[float])-> float
-
-        raise NotImplementedError
-
-    @abc.abstractmethod
-    def suggest_uniform(self, name, low, high):
-        # type: (str, float, float) -> float
-
-        raise NotImplementedError
-
-    @abc.abstractmethod
-    def suggest_loguniform(self, name, low, high):
-        # type: (str, float, float) -> float
-
-        raise NotImplementedError
-
-    @abc.abstractmethod
-    def suggest_discrete_uniform(self, name, low, high, q):
-        # type: (str, float, float, float) -> float
-
-        raise NotImplementedError
-
-    @abc.abstractmethod
-    def suggest_int(self, name, low, high, step=1):
-        # type: (str, int, int, int) -> int
-
-        raise NotImplementedError
-
-    @abc.abstractmethod
-    def suggest_categorical(self, name, choices):
-        # type: (str, Sequence[CategoricalChoiceType]) -> CategoricalChoiceType
-
-        raise NotImplementedError
-
-    @abc.abstractmethod
-    def report(self, value, step):
-        # type: (float, int) -> None
-
-        raise NotImplementedError
-
-    @abc.abstractmethod
-    def should_prune(self, step=None):
-        # type: (Optional[int]) -> bool
-
-        raise NotImplementedError
-
-    @abc.abstractmethod
-    def set_user_attr(self, key, value):
-        # type: (str, Any) -> None
-
-        raise NotImplementedError
-
-    @abc.abstractmethod
-    def set_system_attr(self, key, value):
-        # type: (str, Any) -> None
-
-        raise NotImplementedError
-
-    @property
-    @abc.abstractmethod
-    def params(self):
-        # type: () -> Dict[str, Any]
-
-        raise NotImplementedError
-
-    @property
-    @abc.abstractmethod
-    def distributions(self):
-        # type: () -> Dict[str, BaseDistribution]
-
-        raise NotImplementedError
-
-    @property
-    @abc.abstractmethod
-    def user_attrs(self):
-        # type: () -> Dict[str, Any]
-
-        raise NotImplementedError
-
-    @property
-    @abc.abstractmethod
-    def system_attrs(self):
-        # type: () -> Dict[str, Any]
-
-        raise NotImplementedError
-
-    @property
-    @abc.abstractmethod
-    def datetime_start(self):
-        # type: () -> Optional[datetime.datetime]
-
-        raise NotImplementedError
-
-    @property
-    def number(self) -> int:
-
-        raise NotImplementedError
 
 
 class Trial(BaseTrial):
@@ -1166,200 +818,3 @@ class Trial(BaseTrial):
         self.logger.warning(message)
 
         return self.study._study_id
-
-
-class FixedTrial(BaseTrial):
-    """A trial class which suggests a fixed value for each parameter.
-
-    This object has the same methods as :class:`~optuna.trial.Trial`, and it suggests pre-defined
-    parameter values. The parameter values can be determined at the construction of the
-    :class:`~optuna.trial.FixedTrial` object. In contrast to :class:`~optuna.trial.Trial`,
-    :class:`~optuna.trial.FixedTrial` does not depend on :class:`~optuna.study.Study`, and it is
-    useful for deploying optimization results.
-
-    Example:
-
-        Evaluate an objective function with parameter values given by a user.
-
-        .. testcode::
-
-            import optuna
-
-            def objective(trial):
-                x = trial.suggest_uniform('x', -100, 100)
-                y = trial.suggest_categorical('y', [-1, 0, 1])
-                return x ** 2 + y
-
-            assert objective(optuna.trial.FixedTrial({'x': 1, 'y': 0})) == 1
-
-
-    .. note::
-        Please refer to :class:`~optuna.trial.Trial` for details of methods and properties.
-
-    Args:
-        params:
-            A dictionary containing all parameters.
-        number:
-            A trial number. Defaults to ``0``.
-
-    """
-
-    def __init__(self, params, number=0):
-        # type: (Dict[str, Any], int) -> None
-
-        self._params = params
-        self._suggested_params = {}  # type: Dict[str, Any]
-        self._distributions = {}  # type: Dict[str, BaseDistribution]
-        self._user_attrs = {}  # type: Dict[str, Any]
-        self._system_attrs = {}  # type: Dict[str, Any]
-        self._datetime_start = datetime.datetime.now()
-        self._number = number
-
-    def suggest_float(self, name, low, high, *, log=False, step=None):
-        # type: (str, float, float, bool, Optional[float]) -> float
-
-        if step is not None:
-            if log:
-                raise NotImplementedError(
-                    "The parameter `step` is not supported when `log` is True."
-                )
-            else:
-                return self._suggest(
-                    name, distributions.DiscreteUniformDistribution(low=low, high=high, q=step)
-                )
-        else:
-            if log:
-                return self._suggest(
-                    name, distributions.LogUniformDistribution(low=low, high=high)
-                )
-            else:
-                return self._suggest(name, distributions.UniformDistribution(low=low, high=high))
-
-    def suggest_uniform(self, name, low, high):
-        # type: (str, float, float) -> float
-
-        return self._suggest(name, distributions.UniformDistribution(low=low, high=high))
-
-    def suggest_loguniform(self, name, low, high):
-        # type: (str, float, float) -> float
-
-        return self._suggest(name, distributions.LogUniformDistribution(low=low, high=high))
-
-    def suggest_discrete_uniform(self, name, low, high, q):
-        # type: (str, float, float, float) -> float
-
-        high = _adjust_discrete_uniform_high(name, low, high, q)
-        discrete = distributions.DiscreteUniformDistribution(low=low, high=high, q=q)
-        return self._suggest(name, discrete)
-
-    def suggest_int(self, name, low, high, step=1):
-        # type: (str, int, int, int) -> int
-        sample = self._suggest(
-            name, distributions.IntUniformDistribution(low=low, high=high, step=step)
-        )
-        return int(sample)
-
-    def suggest_categorical(self, name, choices):
-        # type: (str, Sequence[CategoricalChoiceType]) -> CategoricalChoiceType
-
-        choices = tuple(choices)
-        return self._suggest(name, distributions.CategoricalDistribution(choices=choices))
-
-    def _suggest(self, name, distribution):
-        # type: (str, BaseDistribution) -> Any
-
-        if name not in self._params:
-            raise ValueError(
-                "The value of the parameter '{}' is not found. Please set it at "
-                "the construction of the FixedTrial object.".format(name)
-            )
-
-        value = self._params[name]
-        param_value_in_internal_repr = distribution.to_internal_repr(value)
-        if not distribution._contains(param_value_in_internal_repr):
-            raise ValueError(
-                "The value {} of the parameter '{}' is out of "
-                "the range of the distribution {}.".format(value, name, distribution)
-            )
-
-        if name in self._distributions:
-            distributions.check_distribution_compatibility(self._distributions[name], distribution)
-
-        self._suggested_params[name] = value
-        self._distributions[name] = distribution
-
-        return value
-
-    def report(self, value, step):
-        # type: (float, int) -> None
-
-        pass
-
-    def should_prune(self, step=None):
-        # type: (Optional[int]) -> bool
-
-        return False
-
-    def set_user_attr(self, key, value):
-        # type: (str, Any) -> None
-
-        self._user_attrs[key] = value
-
-    def set_system_attr(self, key, value):
-        # type: (str, Any) -> None
-
-        self._system_attrs[key] = value
-
-    @property
-    def params(self):
-        # type: () -> Dict[str, Any]
-
-        return self._suggested_params
-
-    @property
-    def distributions(self):
-        # type: () -> Dict[str, BaseDistribution]
-
-        return self._distributions
-
-    @property
-    def user_attrs(self):
-        # type: () -> Dict[str, Any]
-
-        return self._user_attrs
-
-    @property
-    def system_attrs(self):
-        # type: () -> Dict[str, Any]
-
-        return self._system_attrs
-
-    @property
-    def datetime_start(self):
-        # type: () -> Optional[datetime.datetime]
-
-        return self._datetime_start
-
-    @property
-    def number(self) -> int:
-
-        return self._number
-
-
-def _adjust_discrete_uniform_high(name, low, high, q):
-    # type: (str, float, float, float) -> float
-
-    d_high = decimal.Decimal(str(high))
-    d_low = decimal.Decimal(str(low))
-    d_q = decimal.Decimal(str(q))
-
-    d_r = d_high - d_low
-
-    if d_r % d_q != decimal.Decimal("0"):
-        high = float((d_r // d_q) * d_q + d_low)
-        _logger.warning(
-            "The range of parameter `{}` is not divisible by `q`, and is "
-            "replaced by [{}, {}].".format(name, low, high)
-        )
-
-    return high
